+++ src/extension/alCreateProcedureCA.ts
@@ -5,10 +5,6 @@ import { ALProcedureCallParser } from './alProcedureCallParser';
 import { ALProcedureSourceCodeCreator } from './alProcedureSourceCodeCreator';
 import { SupportedDiagnosticCodes } from './supportedDiagnosticCodes';
 import { ALSourceCodeHandler } from './alSourceCodeHandler';
-import { RenameMgt } from './checkRename';
-import { ALCodeOutlineExtension } from './devToolsExtensionContext';
-import { ToolsGetSyntaxTreeSymbolsRequest } from './ToolsGetSyntaxTreeSymbolsRequest';
-import { ToolsGetSyntaxTreeRequest } from './toolsGetSyntaxTreeRequest';
 import { SyntaxTree } from './AL Code Outline/syntaxTree';
 import { FullSyntaxTreeNodeKind } from './AL Code Outline Ext/fullSyntaxTreeNodeKind';
 import { ALFullSyntaxTreeNode } from './AL Code Outline/alFullSyntaxTreeNode';
