+++ src/extension/renameMgt.ts
